finalized restricting messages for find method
