+++ backend/src/services/messages/messages.hooks.js
@@ -2,14 +2,12 @@ const { authenticate } = require('@feathersjs/authentication').hooks;
 
 const populateUser = require('../../hooks/populate-user');
 const processMessage = require('../../hooks/process-message');
-const filterMessages = require('../../hooks/filter-messages');
+const restrictMessages = require('../../hooks/restrict-messages');
 
 module.exports = {
   before: {
     all: [ authenticate('jwt') ],
-    find: [ 
-      filterMessages()
-    ],
+    find: [restrictMessages()],
     get: [],
     create: [processMessage()],
     update: [],
